requirements.txt
bcrypt
azure-search-documents
azure-core
fuzzywuzzy
psycopg2
streamlit
python-dotenv
fastapi
PyPDF2
pydantic
langchain
langchain-openai
beautifulsoup4
requests
tiktoken
python-Levenshtein
streamlit-webrtc
azure-cognitiveservices-speech
streamlit-autorefresh